split glue file into individual files
